Extract API calls to a custom hook for simplicity
